Move FileTypeManager out of editing module
