+++ indico/modules/events/client/js/reviewing/file_types/FileTypeManager.tsx
