+++ indico/modules/events/client/js/reviewing/file_types/FileTypeModal.jsx
