+++ indico/modules/events/client/js/reviewing/file_types/index.js
@@ -1,0 +1,8 @@
+// This file is part of Indico.
+// Copyright (C) 2002 - 2026 CERN
+//
+// Indico is free software; you can redistribute it and/or
+// modify it under the terms of the MIT License; see the
+// LICENSE file for more details.
+
+export {default as FileTypeManager} from './FileTypeManager';
